Created a function
This functions takes an existing null model, applies a transformation
on the outcome, and fits another null model using the transformed
outcome and all other parameters as in the first null model.
Another function in this file calculates the average variance across
all observations in a given group of observations.